requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

fastapi
scikit-learn
umap-learn
opencv-python
numpy
uvicorn
Pillow
python-multipart
pymongo
python-dotenv
torch
torchvision
cupy